# AlunoAPI/TestePraticoUpBase/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Teste Pratico UpBase - AlunoAPI",
    "description": "API RESTful criada como requisito avaliativo da fase prática do processo seletivo para vaga de estágio .NET na UpBase. Esta API conta com recursos CRUD, integração com banco de dados MySQL, e recursos extra como senha criptografada e critérios que impedem o cadastro de usuário com dados duplicados.",
    "version": "0.0.1",
    "contact": {
      "name": "Suporte",
      "email": "henriqueweege@gmail.com",
      "url": "https://www.linkedin.com/in/henrique-weege-baa287166/"
    },
    "license": {
      "name": "Lincença: GPLv3",
      "url": "https://www.gnu.org/licenses/gpl-3.0.html"
    }
  },
  "paths": {
    "/Aluno": {
      "post": {
        "summary": "Esse método permite cadastrar um novo aluno no bando de dados.",
        "tags": [
          "Aluno API CRUD"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlunoDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlunoDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlunoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Aluno adicionado com sucesso."
          },
          "400": {
            "description": "Todos os campos devem estar preenchidos."
          }
        }
      },
      "get": {
        "summary": "Esse método permite recuperar os dados de todos os alunos.",
        "tags": [
          "Aluno API CRUD"
        ],
        "responses": {
          "200": {
            "description": "Alunos recuperados com sucesso."
          },
          "404": {
            "description": "Não há alunos cadastrados."
          }
        }
      }
    },
    "/Aluno/{id}": {
      "get": {
        "summary": "Esse método permite recuperar os dados do aluno a partir do seu ID.",
        "tags": [
          "Aluno API CRUD"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aluno recuperado com sucesso."
          },
          "404": {
            "description": "ID inválido."
          }
        }
      },
      "put": {
        "summary": "Esse método permite editar o cadastro do aluno a partir do seu ID.",
        "tags": [
          "Aluno API CRUD"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlunoDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlunoDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlunoDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Cadastro alterado com sucesso."
          },
          "404": {
            "description": "ID inválido."
          }
        }
      },
      "delete": {
        "summary": "Esse método permite deletar o cadastro do aluno a partir do seu ID.",
        "tags": [
          "Aluno API CRUD"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Cadastro excluído com sucesso."
          },
          "404": {
            "description": "ID inválido."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateAlunoDto": {
        "required": [
          "email",
          "id",
          "nomeCompleto",
          "nomeDeUsuario",
          "password",
          "rePassword"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nomeCompleto": {
            "type": "string"
          },
          "nomeDeUsuario": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password"
          }
        },
        "additionalProperties": false
      },
      "UpdateAlunoDto": {
        "type": "object",
        "properties": {
          "nomeCompleto": {
            "type": "string",
            "nullable": true
          },
          "nomeDeUsuario": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "format": "password",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}